backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
firebase-admin>=6.6.0
channels>=4.0
daphne>=4.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
